Fix compilations issues

+++ source/blender/draw/engines/eevee/shaders/eevee_deferred_thickness_amend_frag.glsl
@@ -125,7 +125,7 @@ void main()
   gbuffer::Header header = gbuffer::Header::from_data(
       texelFetch(gbuf_header_tx, int3(texel, 0), 0).r);
 
-  int data_layer = header.closure_len();
+  uchar data_layer = header.closure_len();
   float2 data_packed = imageLoad(gbuf_normal_img, int3(texel, data_layer)).rg;
   float gbuffer_thickness = gbuffer::thickness_unpack(data_packed.x);
   if (gbuffer_thickness == 0.0f) {
